+++ src/polymorphic/templatetags/polymorphic_admin_tags.py
@@ -43,6 +43,23 @@ class BreadcrumbScope(Node):
 @register.tag
 def breadcrumb_scope(parser, token):
     """
+    .. templatetag:: breadcrumb_scope
+
     Easily allow the breadcrumb to be generated in the admin change templates.
+
+    The ``{% breadcrumb_scope ... %}`` tag makes sure the ``{{ opts }}`` and
+    ``{{ app_label }}`` values are temporary based on the provided
+    ``{{ base_opts }}``.
+
+    This allows fixing the breadcrumb in admin templates:
+
+    .. code-block:: html+django
+
+        {% extends "admin/change_form.html" %}
+        {% load polymorphic_admin_tags %}
+
+        {% block breadcrumbs %}
+        {% breadcrumb_scope base_opts %}{{ block.super }}{% endbreadcrumb_scope %}
+        {% endblock %}
     """
     return BreadcrumbScope.parse(parser, token)