+++ src/polymorphic/tests/errata/settings.py
@@ -1,0 +1,6 @@
+from ..settings import *
+
+INSTALLED_APPS = [
+    *INSTALLED_APPS,
+    "polymorphic.tests.errata",
+]