+++ src/polymorphic/tests/examples/integrations/extra_views/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class ExtraViewsExampleConfig(AppConfig):
+    name = "polymorphic.tests.examples.integrations.extra_views"
+    label = "extra_views"
+    verbose_name = "django-extra-views Integration Example"