+++ src/polymorphic/utils.py
@@ -1,8 +1,12 @@
+from collections import defaultdict
 from dataclasses import dataclass
 from functools import lru_cache
 
+from django.apps import apps
 from django.contrib.contenttypes.models import ContentType
+from django.core.exceptions import FieldError
 from django.db import DEFAULT_DB_ALIAS, models
+from django.db.models import Q, Subquery
 
 
 @dataclass(frozen=True)
@@ -125,8 +129,16 @@ def route_to_ancestor(model_class, ancestor_model):
     return found_route
 
 
+def is_model_loaded(model):
+    try:
+        apps.get_model(model._meta.app_label, model._meta.model_name)
+        return True
+    except LookupError:
+        return False
+
+
 @lru_cache(maxsize=None)
-def concrete_descendants(model_class):
+def concrete_descendants(model_class, include_proxy=False):
     """
     Get a list of all concrete (non-abstract, non-proxy) descendant model classes in
     tree order with leaf descendants last. Results are cached.
@@ -139,8 +151,9 @@ def concrete_descendants(model_class):
         """Add concrete descendants in tree order (ancestors before descendants)."""
         for sub_cls in model.__subclasses__():
             # Add concrete models in pre-order (parent before children)
-            if not sub_cls._meta.abstract and not sub_cls._meta.proxy:
-                result.append(sub_cls)
+            if not sub_cls._meta.abstract and (include_proxy or not sub_cls._meta.proxy):
+                if is_model_loaded(sub_cls):
+                    result.append(sub_cls)
 
             # Always recurse to find descendants through abstract and proxy models
             add_concrete_descendants(sub_cls, result)
@@ -215,3 +228,62 @@ def prepare_for_copy(obj):
 
     reset_parent_pointers(obj)
     obj._state.adding = True  # Mark as new object
+
+
+def _lazy_ctype(model, using=DEFAULT_DB_ALIAS):
+    """
+    Return the content type id for the given model class if it is in the cache,
+    otherwise return a subquery that can be used to match the content type as part
+    of a larger query. Safe to call before apps are fully loaded.
+
+    :param model: The model class to get the content type for.
+    :return: The content type for the model class.
+    :rtype: int or Subquery
+    """
+    mgr = ContentType.objects.db_manager(using=using)
+    if apps.models_ready and (
+        cid := mgr._cache.get(using, {}).get((model._meta.app_label, model._meta.model_name))
+    ):
+        return cid
+    return Q(app_label=model._meta.app_label) & Q(model=model._meta.model_name)
+
+
+def lazy_ctype(model, using=DEFAULT_DB_ALIAS):
+    ctype = _lazy_ctype(model, using=using)
+    return (
+        ctype
+        if isinstance(ctype, ContentType)
+        else Subquery(ContentType.objects.db_manager(using=using).filter(ctype).values("pk")[:1])
+    )
+
+
+@lru_cache(maxsize=None)
+def _map_queryname_to_class(base_model, qry_name):
+    """Try to match a model name in a query to a model class"""
+    name_map = defaultdict(list)
+    name_map[base_model.__name__.lower()].append(base_model)
+    for cls in concrete_descendants(base_model, include_proxy=True):
+        name_map[cls.__name__.lower()].append(cls)
+
+    matches = name_map.get(qry_name.lower(), [])
+    if len(matches) == 1:
+        return matches[0]
+    elif len(matches) > 1:
+        raise FieldError(
+            f"{qry_name} could refer to any of {[m._meta.label for m in matches]}. In "
+            f"this case, please use the syntax: applabel__ModelName___field"
+        )
+
+    # FIXME: raise a FieldError - upstream code currently relies on this AssertionError
+    # and it will be thrown in legitimate cases because this function ends up being
+    # called on subclasses of the original query model in _get_real_instances. That
+    # code should be refactored to avoiid this.
+    raise AssertionError(f"{qry_name} is not a subclass of {base_model._meta.label}")
+
+
+def _clear_utility_caches():
+    """Clear all lru_cache caches in this module."""
+    get_base_polymorphic_model.cache_clear()
+    route_to_ancestor.cache_clear()
+    concrete_descendants.cache_clear()
+    _map_queryname_to_class.cache_clear()